feat: Handle "synthesized" class/member annotations (#52)
This adds support for "synthesized" class and member annotations.

* It adds a new variant `Synthesized` to `R8Header` to parse these
annotations.
* It adds `is_synthesized` flags to `ClassMapping` and `MemberMapping`
and their corresponding cache types. In the cache the flags have type
`u8` because `bool` would be unsound to transmute.
* It adds an `method_synthesized` flag to `StackFrame` that is set if the
frame's method is flagged as synthesized.
* It adds cache-using variants to the tests in `callback.rs`.
* It bumps the version of the proguard cache format to 2. The version
needs to be increased because records now have an extra field.

With this change, `remap_frame` (both the `mapper` and the `cache`
version) will return exactly the same frames as before, but some of them
might be marked as synthesized. This means it's up to the caller to
decide what to do with synthesized frames.

Closes #48. Closes RUSTPRO-3.